added two functions to the dbquestionhandler class
getQuestionSets() in class DBQuestionHandler
Called from TakeQuiz.java - this will pull all the test sets for display so the user can click on one to take the test.

takeTheQuiz(string) in class DBQuestionHandler
called from TakeQuiz.java - this will take the test set the user clicks on and run a query to pull all the questions to later display.

 i made a lot of notes/comments in this code so you can work on integrating it in. i will continue to help as well.