see previous push notes...
to get to see all the data login, then click on take quiz. This will run the two functions right now just to show functionality. it will first pull all the question sets which we will then be able to display, then once the user clicks on which set they would like to take (set = test), it will be passed to the second function. This second function pulls all the questions that are in that set/test.

right now it logs all of them so you can see the questions, answers, right answer, etc. we can use this to display the data in the view instead of logging.

let me know what you think. i may not have time to work on this tomorrow night, but maybe. let me know how you all are coming along.